% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_data_col}
\alias{clean_data_col}
\title{Apply common data cleaning methods}
\usage{
clean_data_col(.x)
}
\arguments{
\item{.x}{a vector of data of any basic data type}
}
\value{
a vector of clean data
}
\description{
Cleans a vector of data by applying common data cleaning operations.
}
\details{
Any data type can be passed to the function, but some type checking is done
before an operation is performed. Character vectors are stripped of the following
characters: \code{-}
}
\examples{
\donttest{
## Clean a column of character data
v <- c("cat","dog-cat","c-a-t","-catdo-g")
clean_data_col(v)
## [1] "cat"    "dogcat" "cat"    "catdog"
}

}
