% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_data}
\alias{compare_data}
\title{Get names of matching columns}
\usage{
compare_data(data, test_data)
}
\arguments{
\item{data}{theoretical dataset}

\item{test_data}{empirical dataset}
}
\value{
a vector of column names
}
\description{
Return a vector of columns names found in two given data frames.
Currently only names must match for the column to be chosen.
}
\details{
Both names and class should match for the column to be identified as a match.
Currently only names are checked and matching names is case-sensitive.
}
\examples{
\donttest{
a <- data.frame(col_one = c(1,2,3), col_two = c("a","b","c"))
b <- data.frame(col_one = c(1,2,3), col_two = c("d","e","f"))
compare_data(a,b)
c <- data.frame(col_One = c(1,2,3), col_two = c("a","b","c"))
compare_data(a,b)
d <- data.frame(col_one = c(TRUE, FALSE, FALSE), col_two = c("a","b","c"))
compare_data(a,b)
}

}
