% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Create folder structure for sample data}
\usage{
create_folder_structure(name, folder_location = "./data")
}
\arguments{
\item{name}{name of parquet data or partition}

\item{folder_location}{file path to working folder, defaults to \code{dipsim/data}}
}
\description{
Creates a folder structure at \code{location} specified, with same name as parquet data.
if no folder exists at the specified location, it is created.
}
\examples{
\dontrun{
## creates ./data/ido_mcfd and subfolders
create_folder_structure(name = "ido_mcdf") 
## or
create_folder_structure("/20210201/ido_mcdf") 

## returns a file path, fp
fp <- search_parquet_data() 
create_folder_structure(fp)
}

}
