% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_univariate.R
\name{create_univariate}
\alias{create_univariate}
\alias{create_continuous}
\alias{create_discrete}
\title{A title for create_univariate}
\usage{
create_univariate(v, samp_size = 20)

create_continuous(v)

create_discrete(v)
}
\arguments{
\item{v}{vector of values to simulate}

\item{samp_size}{integer representing size of empirical distribution}
}
\description{
Details of create_univariate
}
\section{Functions}{
\itemize{
\item \code{create_continuous()}: A Title for create_continuous

\item \code{create_discrete()}: A Title for create_discrete

}}
\examples{
\dontrun{
create_univariate(rnorm(10))
}

}
