% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_empirical_df.R
\name{generate_empirical}
\alias{generate_empirical}
\title{write_empirical_df}
\usage{
generate_empirical(samp_size = 50, folder_location, name)
}
\arguments{
\item{samp_size}{integer, desired row count of empirical data, defaults to 100}

\item{folder_location}{path to data, relative to the package folder, defaults to ./data}

\item{name}{name of dataset to simulate}
}
\description{
A title for write_empirical_df
}
\details{
Details of write_empirical_df
}
\examples{
\dontrun{
wd <- 'R:/working/users/brobert/mydata'
fp <- search_parquet_data()
generate_testdata(samp_size = 50, folder_location = wd, name = fp)
}
}
