% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testdata.R
\name{generate_testdata}
\alias{generate_testdata}
\title{Generate a simulated dataset}
\usage{
generate_testdata(folder_location = "./data", name, dataset_size = 1000)
}
\arguments{
\item{folder_location}{path to data, relative to the package folder, defaults to ./data}

\item{name}{name of dataset to simulate}

\item{dataset_size}{desired size of the dataset to return, defaults to 1000 rows}
}
\description{
Generate a simulated dataset
}
\details{
Generate a dataset of any given size based on a pdf or pmf stored
at a given folder location.  The pmf or pdf are generated in the
function \code{generate_empirical() } in this package.
}
\examples{
\dontrun{
wd <- 'R:/working/users/brobert/mydata'
fp <- search_parquet_data()
generate_testdata(folder_location = wd, name = fp, dataset_size = 100)
}

}
