% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_theoretical}
\alias{generate_theoretical}
\title{Generate a sample theoretical distribution}
\usage{
generate_theoretical(support, resize = 1e+05, folder_location = "./data")
}
\arguments{
\item{support}{file path to parquet data}

\item{resize}{an integer number of rows to sample, defaults to 100000}

\item{folder_location}{a path to a folder to write data.  defaults to \code{./data}}
}
\value{
sample of a parquet dataset
}
\description{
obtains a sample of a parquet dataset for easier memory handling.
Rows are randomly selected using sample().
}
\examples{
\dontrun{
## Generate a sample of parquet data
wd <- 'R:/working/users/brobert/mydata'
fp <- search_parquet_data()
df <- generate_theoretical(support = fp, resize = 100, folder_location = wd)
}

}
