% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input_data.R
\name{make_input_data}
\alias{make_input_data}
\title{Generate an input an data set}
\usage{
make_input_data(support_fp, resize = 1e+05, folder_location = ".")
}
\arguments{
\item{support_fp}{file path to support data}

\item{resize}{an integer number of rows to sample, defaults to 100000}

\item{folder_location}{a path to a folder to write data.  defaults to \code{.}}
}
\value{
a sample of the support data or the full dataset.
}
\description{
obtains a sample of a parquet dataset for easier memory handling.
Rows are randomly selected using sample(). Setup files and folder structure
to house simulated data is also created.
}
\examples{
\dontrun{
## Generate a sample of parquet data
wd <- 'R:/working/users/brobert/mydata'
fp <- search_parquet_data()
df <- make_input_data(support_fp = fp, resize = 100, folder_location = wd)
}

}
