% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulated_data.R
\name{make_simulated_data}
\alias{make_simulated_data}
\title{Generate a simulated dataset}
\usage{
make_simulated_data(samp_size = 50, folder_location, name, dataset_size = 100)
}
\arguments{
\item{samp_size}{integer, desired row count of base data, defaults to 100}

\item{folder_location}{path to data, relative to the package folder, defaults to ./}

\item{name}{name of dataset to simulate}

\item{dataset_size}{desired size of the dataset to return, defaults to 1000 rows}
}
\value{
a simulated dataset
}
\description{
Generate a simulated dataset
}
\details{
Generate a dataset of any given size based on a pdf or pmf stored
at a given folder location.  The pmf or pdf are generated in the
function \code{generate_empirical() } in this package.
}
\examples{
\dontrun{
wd <- '.'
fp <- search_parquet_data()
generate_empirical(samp_size = 50, folder_location = wd, name = fp, dataset_size = 100)
}
}
