% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testdata.R
\name{read_testdata}
\alias{read_testdata}
\title{Read a dataset into memory}
\usage{
read_testdata(folder_location = "./data", name)
}
\arguments{
\item{folder_location}{path to data, relative to the package folder, defaults to ./data}

\item{name}{name of dataset to simulate}
}
\value{
a simulated dataset
}
\description{
Read a dataset into memory
}
\details{
read a dataset from a given folder location. Metadata are stored along
with the dataset and used in \code{read_csv()} to specify column types and
which columns to read in.
}
\examples{
\dontrun{
wd <- 'R:/working/users/brobert/mydata'
fp <- search_parquet_data()
read_testdata(folder_location = wd, name = fp)
}

}
