% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dataset.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Dynamically plot a dataset}
\usage{
plot_distribution(x, ttl = NULL, ...)
}
\arguments{
\item{x}{column to plot}

\item{ttl}{plot title}

\item{...}{others parameters passed to function}
}
\value{
a ggplot grob with plot
}
\description{
Dynamically plot a dataset
}
\examples{
\dontrun{
plot_distribution(rnorm(100))
}

}
