% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_sim.R
\name{vis_sim}
\alias{vis_sim}
\title{Title}
\usage{
vis_sim(data, test_data, cols)
}
\arguments{
\item{data}{simulated data}

\item{test_data}{original data}

\item{cols}{columns to plot, must be in both data and test_data}
}
\description{
Title
}
\examples{
\dontrun{
vis_sim(df1, df2, c(col1, col2))
}

}
